Allow long modules
and non-toplevel imports.

The former is acceptable for us because we write very long docstrings
since we are doctesting a lot.

The latter is a matter of style. While it is discouraged by PEP8, people
tend to disagree about this one. It can be beneficial not to load
everything from the start, e.g., for startup performance.